Suhani Jain/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Ollama
langchain-ollama>=0.1.0

# Vector Store
chromadb>=0.4.24

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# PDF Loader
pypdf>=3.17.0

# Deepgram SDK
#deepgram-sdk>=3.5.0
deepgram-sdk==3.3.0

# Audio
sounddevice>=0.4.6
scipy>=1.11.0

# Env
python-dotenv>=1.0.0
aiohttp>=3.9.0